+++ changelog.md
@@ -10,6 +10,7 @@
 ---
 
 * [Changelog](#changelog)
+  * [Releases v1.5.0](#releases-v150)
   * [Releases v1.4.1](#releases-v141)
   * [Releases v1.4.0](#releases-v140)
   * [Releases v1.3.0](#releases-v130)
@@ -23,6 +24,12 @@
 ---
 
 ## Changelog
+
+### Releases v1.5.0
+
+1. Fix `multiple-definitions` linker error. Drop `src_cpp` and `src_h` directories
+2. Add feature to select among highest, medium or lowest accuracy for Timers for shortest, medium or longest time
+3. Fix reattachInterrupt() bug. Check [bugfix: reattachInterrupt() pass wrong frequency value to setFrequency() #19](https://github.com/khoih-prog/ESP8266TimerInterrupt/pull/19)
 
 ### Releases v1.4.1